reorganize react routing and add campaign show page
