+++ app/javascript/react/campaigns/CampaignShowContainer.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const CampaignShowContainer = props => {
+  return(
+    <div className="panel">
+      hi
+    </div>
+  )
+}
+
+export default CampaignShowContainer
