+++ app/views/layouts/application.html.erb
@@ -10,14 +10,14 @@
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Quattrocento+Sans&display=swap">
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.2/css/all.css" integrity="sha384-oS3vJWv+0UjzBfQzYUhtDYW+Pj2yciDJxpsK1OYPAYjqT085Qq/1cq5FLXAZQ7Ay" crossorigin="anonymous">
 
-    <%= stylesheet_link_tag    'application', media: 'all', 'data-turbolinks-track': 'reload' %>
+    <%= stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track': 'reload' %>
     <%= javascript_include_tag "vendor/modernizr" %>
     <%= favicon_link_tag 'favicon.ico' %>
   </head>
 
   <body>
     <main id="main">
-      <%- if Rails.env.development? -%>
+      <%- if Rails.env.development? && false -%>
         <div class="dev-links">
           <%= link_to "pry", "/dev/pry" %>
           <%= link_to "user 1", "/dev/login/1" %>
